HW15

+++ CSharpSnake/CSharpSnake/Vector2.cs
@@ -8,8 +8,8 @@ namespace CSharpSnake
 {
     public struct Vector2
     {
-        public float x; public float y;
-        public Vector2(float x, float y)
+        public int x; public int y;
+        public Vector2(int x, int y)
         {
             this.x = x;
             this.y = y;
